requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arrow==1.3.0
art==6.5
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
curl_cffi==0.13.0
dataclasses-json==0.6.7
diff-match-patch==20241021
distro==1.9.0
Faker==37.11.0
fastapi==0.119.0
fastuuid==0.13.5
filelock==3.20.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
guardrails-ai==0.6.7
guardrails-api-client==0.4.0
guardrails_hub_types==0.0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.34
lark==1.3.0
lazy_imports==1.0.1
litellm==1.78.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.17.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.32.1
pycparser==2.23
pydantic==2.12.1
pydantic-settings==2.11.0
pydantic_core==2.41.3
pydash==8.0.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rstr==3.2.2
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.15.4
types-python-dateutil==2.9.0.20251008
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
webcolors==24.11.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
